requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Caching
redis==5.0.1
aioredis==2.0.1

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0

# AI Integration
anthropic==0.7.8
openai==1.3.7

# Notion Integration
notion-client==2.2.1

# Background Jobs
celery==5.3.4

# Search Engine
elasticsearch==8.11.1

# Data Processing
pandas==2.1.4
numpy==1.24.4

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1